BDOG-860 move backward compatibility type alias to avoid overlapping package object
