+++ bootstrap-common-play-26/src/main/scala/uk/gov/hmrc/play/bootstrap/http/HttpClient.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.play.bootstrap
+package uk.gov.hmrc.play.bootstrap.http
 
 import akka.actor.ActorSystem
 import com.typesafe.config.Config
@@ -26,29 +26,21 @@ import uk.gov.hmrc.play.audit.http.HttpAuditing
 import uk.gov.hmrc.play.audit.http.connector.AuditConnector
 import uk.gov.hmrc.play.http.ws._
 
-package object http {
-  @deprecated("Use uk.gov.hmrc.http.HttpClient instead", "2.15.0")
-  type HttpClient = uk.gov.hmrc.http.HttpClient
+@Singleton
+class DefaultHttpClient @Inject()(
+  config: Configuration,
+  val httpAuditing: HttpAuditing,
+  override val wsClient: WSClient,
+  override protected val actorSystem: ActorSystem
+) extends uk.gov.hmrc.http.HttpClient
+     with WSHttp {
+
+  override lazy val configuration: Option[Config] = Option(config.underlying)
+
+  override val hooks: Seq[HttpHook] = Seq(httpAuditing.AuditingHook)
 }
 
-package http {
-
-  @Singleton
-  class DefaultHttpClient @Inject()(
-    config: Configuration,
-    val httpAuditing: HttpAuditing,
-    override val wsClient: WSClient,
-    override protected val actorSystem: ActorSystem
-  ) extends uk.gov.hmrc.http.HttpClient
-       with WSHttp {
-
-    override lazy val configuration: Option[Config] = Option(config.underlying)
-
-    override val hooks: Seq[HttpHook] = Seq(httpAuditing.AuditingHook)
-  }
-
-  class DefaultHttpAuditing @Inject()(
-    val auditConnector: AuditConnector,
-    @Named("appName") val appName: String
-  ) extends HttpAuditing
-}
+class DefaultHttpAuditing @Inject()(
+  val auditConnector: AuditConnector,
+  @Named("appName") val appName: String
+) extends HttpAuditing
